Add files via upload
net (8).js was not inspiring, but it seemed to do the right thing.  

i wanted to see what it would take to graft another layer without destroying the precious cobwebs it depends on.

net (10).js the reserialized edition of net8.js after the client loads it, showing that at least the js is doing what was intended if not the mechanics.